+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1764,7 +1764,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M25CQYAQGRBEF4Y4D2ZDRZ9F"),
+  _id: z.string().default("tool_01M25T8BP4T92V7HWS0A0VBNST"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -1959,7 +1959,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M25CQYAQVGCA1F31EG33B4VH"),
+  _id: z.string().default("tool_01M25T8BP3210FJYQKP18D8Z9D"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2054,7 +2054,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M25CQYAP4PSKYPZ7ECG3R80P"),
+  _id: z.string().default("tool_01M25T8BP344DGQS6BZYFQSW7N"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2157,7 +2157,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M25CQYAPFSHNNK3N5E7YFZ1Z"),
+  _id: z.string().default("tool_01M25T8BP2GDH3SS88E31G4KY5"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),